+++ src/rrt/2dplane/ObstacleGrid.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 namespace RRT {
 
-ObstacleGrid::ObstacleGrid(float width, float height, int discretizedWidth,
+ObstacleGrid::ObstacleGrid(double width, double height, int discretizedWidth,
                            int discretizedHeight) {
     _width = width;
     _height = height;
@@ -22,16 +22,16 @@ ObstacleGrid::ObstacleGrid(float width, float height, int discretizedWidth,
 
 ObstacleGrid::~ObstacleGrid() { free(_obstacles); }
 
-Vector2i ObstacleGrid::gridSquareForLocation(const Vector2f& loc) const {
+Vector2i ObstacleGrid::gridSquareForLocation(const Vector2d& loc) const {
     return Vector2i(loc.x() / width() * discretizedWidth(),
                     loc.y() / height() * discretizedHeight());
 }
 
-float ObstacleGrid::nearestObstacleDist(const Vector2f& state,
-                                        float maxDist) const {
+double ObstacleGrid::nearestObstacleDist(const Vector2d& state,
+                                        double maxDist) const {
     // x and y are the indices of the cell that state is located in
-    float x = (state.x() / (_width / _discretizedWidth));
-    float y = (state.y() / (_height / _discretizedHeight));
+    double x = (state.x() / (_width / _discretizedWidth));
+    double y = (state.y() / (_height / _discretizedHeight));
     int xSearchRad = maxDist * _discretizedWidth / _width;
     int ySearchRad = maxDist * _discretizedHeight / _height;
     // here we loop through the cells around (x,y) to find the minimum distance
@@ -41,9 +41,9 @@ float ObstacleGrid::nearestObstacleDist(const Vector2f& state,
         for (int j = y - ySearchRad; j <= y + ySearchRad; j++) {
             bool obs = obstacleAt(i, j);
             if (obs) {
-                float xDist = (x - i) * _width / _discretizedWidth;
-                float yDist = (y - j) * _height / _discretizedHeight;
-                float dist = sqrtf(powf(xDist, 2) + powf(yDist, 2));
+                double xDist = (x - i) * _width / _discretizedWidth;
+                double yDist = (y - j) * _height / _discretizedHeight;
+                double dist = sqrtf(powf(xDist, 2) + powf(yDist, 2));
                 if (dist < maxDist) {
                     maxDist = dist;
                 }
@@ -88,8 +88,8 @@ int ObstacleGrid::discretizedWidth() const { return _discretizedWidth; }
 
 int ObstacleGrid::discretizedHeight() const { return _discretizedHeight; }
 
-float ObstacleGrid::width() const { return _width; }
+double ObstacleGrid::width() const { return _width; }
 
-float ObstacleGrid::height() const { return _height; }
+double ObstacleGrid::height() const { return _height; }
 
 }  // namespace RRT
