+++ src/rrt/StateSpace.hpp
@@ -29,7 +29,7 @@ public:
      * that.
      */
     virtual T intermediateState(const T& source, const T& target,
-                                float stepSize) const = 0;
+                                double stepSize) const = 0;
 
     /**
      * An overloaded version designed for use in adaptive stepsize control.
@@ -44,7 +44,7 @@ public:
      * @return A state in the direction of @target from @source.state()
      */
     virtual T intermediateState(const T& source, const T& target,
-                                float minStepSize, float maxStepSize) const = 0;
+                                double minStepSize, double maxStepSize) const = 0;
 
     /**
      * @brief Calculate the distance between two states
@@ -77,8 +77,8 @@ public:
     virtual bool transitionValid(const T& from, const T& to) const = 0;
 
 protected:
-    float _minStepSize;
-    float _maxStepSize;
+    double _minStepSize;
+    double _maxStepSize;
 };
 
 }  // namespace RRT
